+++ tests/testthat/test-extend_to_useful.R
@@ -15,14 +15,14 @@ test_that("C matrices are created correctly", {
 
 
 test_that("eta_fu and phi_u vectors are created correctly", {
-  testing_setup <- SEAPSUTWorkflow:::set_up_for_testing(how_far = SEAPSUTWorkflow::target_names$EtaPhivecs)
+  testing_setup <- SEAPSUTWorkflow:::set_up_for_testing(how_far = SEAPSUTWorkflow::target_names$EtafuPhiuvecs)
   tryCatch({
     drake::make(testing_setup$plan, cache = testing_setup$temp_cache, verbose = 0)
 
     # Check that the eta_fu and phi_u were added to the WithEtaPhi_target data frame target
-    EtaPhi_target <- readd(SEAPSUTWorkflow::target_names$EtaPhivecs, character_only = TRUE, path = testing_setup$cache_path)
-    expect_true(IEATools::template_cols$eta_fu %in% colnames(EtaPhi_target))
-    expect_true(IEATools::template_cols$phi_u %in% colnames(EtaPhi_target))
+    EtafuPhiu_target <- readd(SEAPSUTWorkflow::target_names$EtafuPhiuvecs, character_only = TRUE, path = testing_setup$cache_path)
+    expect_true(IEATools::template_cols$eta_fu %in% colnames(EtafuPhiu_target))
+    expect_true(IEATools::template_cols$phi_u %in% colnames(EtafuPhiu_target))
   },
   finally = {
     SEAPSUTWorkflow:::clean_up_after_testing(testing_setup)