+++ NEWS.md
@@ -1,3 +1,9 @@
+---
+title: "Release notes for `SEAPSUTWorkflow`"
+output: html_document
+---
+
+
 * Change name of constant: `row_code` --> `region_code`.