+++ NEWS.md
@@ -4,7 +4,15 @@ output: html_document
 ---
 
 
+* Fixed a bug where the primary-to-final exergy-to-energy ratio
+  (phi) didn't get row and column types.
+* New targets:
+    * `phi_constants_path` gives path to a phi constants file.
+    * `Phi_constants` is a table of reasonable constant values of phi (the energy-to-exergy) ratio.
 * Change name of constant: `row_code` --> `region_code`.
+* New tests for new features.
+  * Now up to 306 tests, all passing.
+  * Test coverage remains at 100%.
 
 
 # SEAPSUTWorkflow 0.2.3 (2021-08-20) [![DOI](https://zenodo.org/badge/265032888.svg)](https://zenodo.org/badge/latestdoi/265032888)